Applying review recommendations.
